new smw backgrounds for donut plains
new bgs for donut plains. bg30 is used both horizontally and vertically but the looping is weirdly different so i made both a horizontal and vertical version.
also bg_ghost and bg_ghost_priority is the ghost house entrance/exit, thought itd be easier/better to have it be a bg asset than akwardly clumped in the tileset (it has a lot of different tiles that make it up).